Fix clippy lint
The parameter to `unwrap_or` is always evaluated, even if the value is
not needed, therefore it should only be used with values that are cheap
to create (or already created). `unwrap_or_else` takes a closure that
returns the default value and is only executed if the value is actually
needed.